new todo functions to support editing items, and new method of adding items
(1) replaceToDo: When editing a ToDo, allow in-place replacement of complete ToDo in ToDo array (rather than updating a single value as with updateToDo). (2) addEmptyToDo: When adding a ToDo, create an empty ToDo card to be edited with new information. (3) addToDo: New ToDos are added to the front of the ToDo array so their card appears at the top of the scroll window.